# backend/src/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Splitsvilla AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contestants:
    get:
      operationId: listContestants
      tags: [contestants]
      summary: List all contestants
      responses:
        "200":
          description: List of contestants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contestant"
    post:
      operationId: createContestant
      tags: [contestants]
      summary: Add a new contestant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContestantBody"
      responses:
        "201":
          description: Created contestant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contestant"
  /contestants/{id}:
    get:
      operationId: getContestant
      tags: [contestants]
      summary: Get a contestant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contestant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contestant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateContestant
      tags: [contestants]
      summary: Update a contestant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContestantBody"
      responses:
        "200":
          description: Updated contestant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contestant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContestant
      tags: [contestants]
      summary: Remove a contestant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /compatibility/analyze:
    post:
      operationId: analyzeCompatibility
      tags: [compatibility]
      summary: AI-powered compatibility analysis between two contestants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompatibilityAnalysisBody"
      responses:
        "200":
          description: Compatibility analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompatibilityResult"
  /compatibility/oracle:
    post:
      operationId: oraclePrediction
      tags: [compatibility]
      summary: Oracle AI prediction for a split/couple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OraclePredictionBody"
      responses:
        "200":
          description: Oracle prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OraclePrediction"
  /compatibility/ml-score:
    post:
      operationId: mlCompatibilityScore
      tags: [compatibility]
      summary: ML-based compatibility score between two contestants
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MlScoreBody"
      responses:
        "200":
          description: ML compatibility score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MlScoreResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Contestant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        hometown:
          type: string
        bio:
          type: string
        zodiacSign:
          type: string
        personality:
          type: string
          description: "Comma-separated personality traits e.g. outgoing,adventurous,loyal"
        interests:
          type: string
          description: "Comma-separated interests"
        status:
          type: string
          enum: [active, dumped, winner]
        gender:
          type: string
          enum: [male, female, other]
        currentSplitId:
          type: integer
          nullable: true
        compatibilityScore:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - age
        - hometown
        - bio
        - zodiacSign
        - personality
        - interests
        - status
        - gender
        - createdAt
    CreateContestantBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        hometown:
          type: string
        bio:
          type: string
        zodiacSign:
          type: string
        personality:
          type: string
        interests:
          type: string
        gender:
          type: string
          enum: [male, female, other]
      required:
        - name
        - age
        - hometown
        - bio
        - zodiacSign
        - personality
        - interests
        - gender
...
    CompatibilityAnalysisBody:
      type: object
      properties:
        contestant1Id:
          type: integer
        contestant2Id:
          type: integer
      required:
        - contestant1Id
        - contestant2Id
    CompatibilityResult:
      type: object
      properties:
        contestant1Name:
          type: string
        contestant2Name:
          type: string
        overallScore:
          type: number
          description: "Score from 0 to 100"
        personalityMatch:
          type: number
        interestMatch:
          type: number
        zodiacMatch:
          type: number
        analysis:
          type: string
        strengths:
          type: array
          items:
            type: string
        challenges:
          type: array
          items:
            type: string
        recommendation:
          type: string
      required:
        - contestant1Name
        - contestant2Name
        - overallScore
        - personalityMatch
        - interestMatch
        - zodiacMatch
        - analysis
        - strengths
        - challenges
        - recommendation
    OraclePredictionBody:
      type: object
      properties:
        contestant1Id:
          type: integer
        contestant2Id:
          type: integer
      required:
        - contestant1Id
        - contestant2Id
    OraclePrediction:
      type: object
      properties:
        prediction:
          type: string
        verdict:
          type: string
          enum: [strong_match, possible_match, uncertain, unlikely_match]
        confidence:
          type: number
        mysticalMessage:
          type: string
        advice:
          type: string
      required:
        - prediction
        - verdict
        - confidence
        - mysticalMessage
        - advice
    MlScoreBody:
      type: object
      properties:
        contestant1Id:
          type: integer
        contestant2Id:
          type: integer
      required:
        - contestant1Id
        - contestant2Id
    MlScoreResult:
      type: object
      properties:
        score:
          type: number
        breakdown:
          type: object
          properties:
            zodiacCompatibility:
              type: number
            personalityOverlap:
              type: number
            interestSimilarity:
              type: number
            ageCompatibility:
              type: number
          required:
            - zodiacCompatibility
            - personalityOverlap
            - interestSimilarity
            - ageCompatibility
        label:
          type: string
          enum: [excellent, good, moderate, poor]
      required:
        - score
        - breakdown
        - label
...