# backend/src/lib/api-zod/src/generated/types/compatibilityResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Splitsvilla AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompatibilityResult {
  contestant1Name: string;
  contestant2Name: string;
  /** Score from 0 to 100 */
  overallScore: number;
  personalityMatch: number;
  interestMatch: number;
  zodiacMatch: number;
  analysis: string;
  strengths: string[];
  challenges: string[];
  recommendation: string;
}